+++ android/src/main/kotlin/net/apexteam/ton/TonPlugin.kt
@@ -1,15 +1,18 @@
 package net.apexteam.ton
 
-import androidx.annotation.NonNull
-
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import org.ton.mnemonic.Mnemonic
+import kotlin.coroutines.CoroutineContext
 
 /** TonPlugin */
-class TonPlugin: FlutterPlugin, MethodCallHandler {
+class TonPlugin: FlutterPlugin, MethodCallHandler, CoroutineScope {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
@@ -22,14 +25,31 @@ class TonPlugin: FlutterPlugin, MethodCallHandler {
   }
 
   override fun onMethodCall(call: MethodCall, result: Result) {
-    if (call.method == "getPlatformVersion") {
-      result.success("Android ${android.os.Build.VERSION.RELEASE}")
-    } else {
-      result.notImplemented()
+    when (call.method) {
+        "getPlatformVersion" -> {
+          result.success("Android ${android.os.Build.VERSION.RELEASE}")
+        }
+        "generateRandomMnemonic" -> {
+            val password = call.argument<String?>("password")
+          launch {
+              val mnemonic = if(password == null) {
+                  Mnemonic.generate()
+              } else {
+                  Mnemonic.generate(password= password)
+              }
+            result.success(mnemonic)
+          }
+        }
+        else -> {
+          result.notImplemented()
+        }
     }
   }
 
   override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)
   }
+
+  override val coroutineContext: CoroutineContext
+    get() = Dispatchers.Main
 }