+++ lib/ton.dart
@@ -5,4 +5,8 @@ class Ton {
   Future<String?> getPlatformVersion() {
     return TonPlatform.instance.getPlatformVersion();
   }
+
+  Future<List<String>?> randomMnemonic({String? password}) {
+    return TonPlatform.instance.generateRandomMnemonic(password: password);
+  }
 }